+++ src/__tests__/upload-client.test.ts
@@ -6,7 +6,7 @@ import http, { type IncomingMessage, type ServerResponse } from 'node:http';
 import os from 'node:os';
 import path from 'node:path';
 import { once } from 'node:events';
-import { uploadArtifact } from '../upload-client.ts';
+import { uploadArtifact } from '../remote/upload-client.ts';
 import type { UploadProgressEvent } from '../upload-progress.ts';
 import { runCmdSync, withCommandExecutorOverride } from '../utils/exec.ts';
