+++ src/cli/cloud-connection-profile.ts
@@ -1,5 +1,5 @@
 import crypto from 'node:crypto';
-import type { RemoteConfigProfile } from '../remote-config-schema.ts';
+import type { RemoteConfigProfile } from '../remote/remote-config-schema.ts';
 import { AppError } from '../kernel/errors.ts';
 import type { CliFlags } from '../utils/cli-flags.ts';
 import type { EnvMap } from '../utils/env-map.ts';
