+++ src/cli/commands/connection-runtime.ts
@@ -1,7 +1,7 @@
 import { resolveDaemonPaths } from '../../daemon/config.ts';
 import { stopReactDevtoolsCompanion } from '../../client-react-devtools-companion.ts';
 import { stopMetroTunnel } from '../../metro/metro.ts';
-import { resolveRemoteConfigProfile } from '../../remote-config.ts';
+import { resolveRemoteConfigProfile } from '../../remote/remote-config.ts';
 import { resolveDevice, type DeviceInfo } from '../../kernel/device.ts';
 import { shouldAgentCdpUseRemoteBridgeUrl } from './agent-cdp.ts';
 import type { MetroBridgeScope } from '../../client-companion-tunnel-contract.ts';
@@ -13,7 +13,7 @@ import {
   writeRemoteConnectionState,
   type RemoteConnectionState,
   type RemoteConnectionRequestMetadata,
-} from '../../remote-connection-state.ts';
+} from '../../remote/remote-connection-state.ts';
 import { profileToCliFlags } from '../../utils/remote-config.ts';
 import type { BatchStep } from '../../client-types.ts';
 import { AppError } from '../../kernel/errors.ts';
