+++ src/cli/commands/proxy.ts
@@ -1,5 +1,5 @@
 import { randomBytes } from 'node:crypto';
-import { createDaemonProxyServer } from '../../daemon-proxy.ts';
+import { createDaemonProxyServer } from '../../remote/daemon-proxy.ts';
 import { buildDaemonHttpBaseUrl } from '../../daemon/http-contract.ts';
 import { ensureDaemon, resolveClientSettings } from '../../daemon-client-lifecycle.ts';
 import { AppError } from '../../kernel/errors.ts';
