+++ src/cli/generated-remote-config.ts
@@ -1,8 +1,11 @@
 import crypto from 'node:crypto';
 import fs from 'node:fs';
 import path from 'node:path';
-import { resolveRemoteConfigProfile } from '../remote-config.ts';
-import type { RemoteConfigProfile, ResolvedRemoteConfigProfile } from '../remote-config-schema.ts';
+import { resolveRemoteConfigProfile } from '../remote/remote-config.ts';
+import type {
+  RemoteConfigProfile,
+  ResolvedRemoteConfigProfile,
+} from '../remote/remote-config-schema.ts';
 import { AppError, asAppError } from '../kernel/errors.ts';
 import type { EnvMap } from '../utils/env-map.ts';
 import type { CliFlags } from '../utils/cli-flags.ts';
