+++ src/remote/daemon-proxy.ts
@@ -2,15 +2,15 @@ import http, { type IncomingMessage, type ServerResponse } from 'node:http';
 import { Readable } from 'node:stream';
 import { pipeline } from 'node:stream/promises';
 import { randomUUID } from 'node:crypto';
-import { AppError, normalizeError } from './kernel/errors.ts';
-import { readNodeHttpRequestBody } from './utils/node-http.ts';
-import { timingSafeStringEqual } from './utils/timing-safe-equal.ts';
+import { AppError, normalizeError } from '../kernel/errors.ts';
+import { readNodeHttpRequestBody } from '../utils/node-http.ts';
+import { timingSafeStringEqual } from '../utils/timing-safe-equal.ts';
 import {
   DAEMON_HTTP_BASE_PATH,
   buildDaemonHttpAuthHeaders,
   buildDaemonHttpUrl,
-} from './daemon/http-contract.ts';
-import { buildDaemonHealthPayload } from './daemon/http-health.ts';
+} from '../daemon/http-contract.ts';
+import { buildDaemonHealthPayload } from '../daemon/http-health.ts';
 
 export type DaemonProxyOptions = {
   upstreamBaseUrl: string;
