+++ src/utils/__tests__/cli-option-schema.test.ts
@@ -8,7 +8,10 @@ import {
   resolveSourceValueDefinition,
 } from '../cli-option-schema.ts';
 import { AppError } from '../../kernel/errors.ts';
-import { REMOTE_CONFIG_FIELD_SPECS, getRemoteConfigEnvNames } from '../../remote-config-schema.ts';
+import {
+  REMOTE_CONFIG_FIELD_SPECS,
+  getRemoteConfigEnvNames,
+} from '../../remote/remote-config-schema.ts';
 
 test('option schema exposes config/env metadata for global options', () => {
   const spec = getOptionSpec('platform');
